+++ base/clients/salesforce/sfc_data.go
@@ -324,7 +324,7 @@ func (cc *SalesforceClient) SearchContact(query string) (*models.SfcContact, *he
 	if len(response.Records) < 1 || response.Records[0].Id == "" {
 		errorMessage := fmt.Sprintf("%s : %s", constants.RequestError, helpers.EmptyResponse)
 		logrus.Error(errorMessage)
-		return nil, &helpers.ErrorResponse{Error: errors.New(errorMessage), StatusCode: err.StatusCode}
+		return nil, &helpers.ErrorResponse{Error: errors.New(errorMessage), StatusCode: http.StatusNotFound}
 	}
 
 	contact := models.SfcContact{